backend/requirements.txt
# SpendX Backend Dependencies

# FastAPI & Server
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-multipart>=0.0.5

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
asyncpg>=0.27.0
alembic>=1.10.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0

# Validation & Settings
pydantic>=2.0.0
pydantic-settings>=2.0.0
email-validator>=2.0.0

# AI
google-generativeai>=0.3.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.24.0

# Development
pytest>=7.0.0
pytest-asyncio>=0.21.0
